+++ weblate/trans/models/component.py
@@ -2047,7 +2047,12 @@ class Component(FastDeleteMixin, models.Model, URLMixin, PathMixin):
 
     def repo_needs_merge(self):
         """Check for unmerged commits from remote repository."""
-        return self.repository.needs_merge()
+        try:
+            return self.repository.needs_merge()
+        except RepositoryException as error:
+            report_error(cause="Could check merge needed")
+            self.add_alert("MergeFailure", error=self.error_text(error))
+            return False
 
     def repo_needs_push(self):
         """Check for something to push to remote repository."""